+++ src/types/api/totalAmountMinted.ts
