+++ src/types/index.ts
@@ -1,12 +1,12 @@
 import {LoginRequest, LoginResponse} from 'types/api/authentication';
 import {CommentReadSerializer, CreateCommentRequest} from 'types/api/comments';
-import {TotalAmountMintedResponse} from 'types/api/currencies';
 import {CurrencyBalance} from 'types/api/currencyBalances';
 import {CreateExchangeOrderRequest} from 'types/api/exchangeOrders';
 import {CreateFollowerRequest, FollowerReadSerializer, GetFollowersParams} from 'types/api/follower';
 import {CreateInvitationRequest} from 'types/api/invitations';
 import {CreateMintRequest, GetMintsParams} from 'types/api/mints';
 import {GetPostsParams, PostReadSerializer} from 'types/api/posts';
+import {TotalAmountMintedResponse} from 'types/api/totalAmountMinted';
 import {GetTradesParams} from 'types/api/trades';
 import {CreateUserRequest, CreateUserResponse, UserReadSerializer} from 'types/api/users';
 import {UserStatsSerializer} from 'types/api/userStats';
